Core game items moved to a separate crate.
