+++ riverraid/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "riverraid"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+log = "0.4.21"
+crossterm = "0.27.0"
+rand = "0.8.5"
+
+[dependencies.uuid]
+version = "1.8.0"
+features = [
+    "v4",       # Lets you generate random UUIDs
+    "fast-rng", # Use a faster (but still sufficiently random) RNG
+]
