+++ riverraid/src/lib.rs
@@ -1,0 +1,19 @@
+mod canvas;
+mod entities;
+mod events;
+mod game;
+mod utilities;
+mod world;
+
+pub use entities::{
+    Bullet, DeathCause, Enemy, Entity, EntityStatus, EntityType, Fuel, Player, PlayerStatus,
+};
+pub use game::Game;
+pub use utilities::event_handler::{
+    EventHandler, IntoEventHandler, IntoTimerEventHandler, LeaveAlone, TimerEventHandler,
+};
+pub use world::{
+    map::{Map, RiverMode},
+    Event, EventTrigger, IntoEvent, IntoEventTrigger, World, WorldEvent, WorldEventTrigger,
+    WorldTimer, WorldTimerData,
+};
