+++ riverraid/src/utilities/stout_ext.rs
